:truck: Move service factory to more general purpose location
We were now duplicating these factories in multiple places and
importing them from weird places.
