+++ src/openforms/prefill/contrib/haalcentraal/tests/test_config_check.py
@@ -10,7 +10,7 @@ from zgw_consumers.models import Service
 from zgw_consumers.test import mock_service_oas_get
 
 from openforms.plugins.exceptions import InvalidPluginConfiguration
-from openforms.registrations.contrib.zgw_apis.tests.factories import ServiceFactory
+from zgw_consumers_ext.factories import ServiceFactory
 
 from ....registry import register
 from ..constants import HaalCentraalVersion
