+++ src/openforms/submissions/tests/form_logic/test_fetching_form_varaiable_values_from_services.py
@@ -13,11 +13,11 @@ from hypothesis import assume, example, given, strategies as st
 from zgw_consumers.constants import APITypes, AuthTypes
 
 from openforms.forms.tests.factories import FormVariableFactory
-from openforms.registrations.contrib.zgw_apis.tests.factories import ServiceFactory
 from openforms.tests.utils import c_profile
 from openforms.variables.constants import DataMappingTypes
 from openforms.variables.tests.factories import ServiceFetchConfigurationFactory
 from openforms.variables.validators import HeaderValidator, ValidationError
+from zgw_consumers_ext.factories import ServiceFactory
 
 from ...logic.service_fetching import perform_service_fetch
