+++ src/openforms/submissions/tests/test_variables/test_fetch_with_logic.py
@@ -8,8 +8,8 @@ from zgw_consumers.constants import APITypes, AuthTypes
 
 from openforms.forms.constants import LogicActionTypes
 from openforms.forms.tests.factories import FormLogicFactory
-from openforms.registrations.contrib.zgw_apis.tests.factories import ServiceFactory
 from openforms.variables.tests.factories import ServiceFetchConfigurationFactory
+from zgw_consumers_ext.factories import ServiceFactory
 
 from ..factories import SubmissionFactory
 from ..mixins import SubmissionsMixin
